# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce API
  version: 0.2.0
  description: API contract for backend and frontend type generation.
servers:
  - url: http://localhost:3000
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/webhooks/events:
    get:
      tags: [admin]
      operationId: listAdminWebhookEvents
      parameters:
        - in: query
          name: provider
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, PROCESSED, DEAD_LETTER, REJECTED]
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Webhook event page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEventPage"
...
  /api/v1/admin/providers/credentials:
    get:
      tags: [admin]
      operationId: listAdminProviderCredentials
      parameters:
        - in: query
          name: provider_type
          schema:
            type: string
            enum: [payment, shipping, tax]
      responses:
        "200":
          description: Provider credential metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderCredentialListResponse"
    post:
      tags: [admin]
      operationId: upsertAdminProviderCredential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderCredentialRequest"
      responses:
        "200":
          description: Stored provider credential metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderCredentialEnvelope"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description: Credential encryption is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/admin/providers/credentials/{id}/rotate:
    post:
      tags: [admin]
      operationId: rotateAdminProviderCredential
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Rotated provider credential metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderCredentialEnvelope"
        "404":
          description: Credential not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description: Credential encryption is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/providers/reconciliation/runs:
    get:
      tags: [admin]
      operationId: listAdminProviderReconciliationRuns
      parameters:
        - in: query
          name: provider_type
          schema:
            type: string
            enum: [payment, shipping, tax]
        - in: query
          name: provider_id
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Provider reconciliation runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderReconciliationRunPage"
    post:
      tags: [admin]
      operationId: createAdminProviderReconciliationRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderReconciliationRunRequest"
      responses:
        "201":
          description: Provider reconciliation run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderReconciliationRunEnvelope"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/providers/reconciliation/runs/{id}:
    get:
      tags: [admin]
      operationId: getAdminProviderReconciliationRun
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Provider reconciliation run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderReconciliationRunEnvelope"
        "404":
          description: Reconciliation run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/webhooks/{provider}:
    post:
      tags: [webhooks]
      operationId: receiveWebhookEvent
      security: []
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookIngestResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    Pagination:
      type: object
      required: [page, limit, total, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
...
    WebhookIngestResponse:
      type: object
      required: [message, event_id, provider_event_id, duplicate]
      properties:
        message:
          type: string
        event_id:
          type: integer
        provider_event_id:
          type: string
        duplicate:
          type: boolean
    WebhookEventRecord:
      type: object
      required:
        - id
        - provider
        - provider_event_id
        - event_type
        - signature_valid
        - payload
        - received_at
        - attempt_count
        - last_error
        - status
      properties:
        id:
          type: integer
        provider:
          type: string
        provider_event_id:
          type: string
        event_type:
          type: string
        signature_valid:
          type: boolean
        payload:
          type: string
        received_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
          nullable: true
        attempt_count:
          type: integer
        last_error:
          type: string
        status:
          type: string
          enum: [PENDING, PROCESSED, DEAD_LETTER, REJECTED]
    WebhookEventPage:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEventRecord"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    ProviderCredentialRequest:
      type: object
      required: [provider_type, provider_id, environment, secret_data]
      properties:
        provider_type:
          type: string
          enum: [payment, shipping, tax]
        provider_id:
          type: string
        environment:
          type: string
          enum: [sandbox, production]
        label:
          type: string
        secret_data:
          type: object
          additionalProperties:
            type: string
        supported_currencies:
          type: array
          items:
            type: string
        settlement_currency:
          type: string
        fx_mode:
          type: string
          enum: [same_currency_only, provider_managed]
    ProviderCredential:
      type: object
      required:
        - id
        - provider_type
        - provider_id
        - environment
        - label
        - key_version
        - supported_currencies
        - fx_mode
        - last_rotated_at
        - updated_at
      properties:
        id:
          type: integer
        provider_type:
          type: string
          enum: [payment, shipping, tax]
        provider_id:
          type: string
        environment:
          type: string
          enum: [sandbox, production]
        label:
          type: string
        key_version:
          type: string
        supported_currencies:
          type: array
          items:
            type: string
        settlement_currency:
          type: string
        fx_mode:
          type: string
          enum: [same_currency_only, provider_managed]
        last_rotated_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProviderCredentialEnvelope:
      type: object
      required: [credential]
      properties:
        credential:
          $ref: "#/components/schemas/ProviderCredential"
    ProviderCredentialListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProviderCredential"
...
    ProviderReconciliationRunRequest:
      type: object
      required: [provider_type, provider_id]
      properties:
        provider_type:
          type: string
          enum: [payment, shipping, tax]
        provider_id:
          type: string
    ProviderReconciliationDrift:
      type: object
      required:
        - id
        - entity_type
        - entity_id
        - provider_reference
        - severity
        - field_name
        - expected_value
        - actual_value
        - message
      properties:
        id:
          type: integer
        entity_type:
          type: string
        entity_id:
          type: integer
        provider_reference:
          type: string
        severity:
          type: string
          enum: [ERROR, WARN]
        field_name:
          type: string
        expected_value:
          type: string
        actual_value:
          type: string
        message:
          type: string
    ProviderReconciliationRun:
      type: object
      required:
        - id
        - provider_type
        - provider_id
        - environment
        - trigger
        - status
        - checked_count
        - drift_count
        - error_count
        - started_at
      properties:
        id:
          type: integer
        provider_type:
          type: string
          enum: [payment, shipping, tax]
        provider_id:
          type: string
        environment:
          type: string
          enum: [sandbox, production]
        trigger:
          type: string
          enum: [MANUAL, SCHEDULED]
        status:
          type: string
          enum: [SUCCEEDED, FAILED]
        checked_count:
          type: integer
        drift_count:
          type: integer
        error_count:
          type: integer
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
        drifts:
          type: array
          items:
            $ref: "#/components/schemas/ProviderReconciliationDrift"
    ProviderReconciliationRunEnvelope:
      type: object
      required: [run]
      properties:
        run:
          $ref: "#/components/schemas/ProviderReconciliationRun"
    ProviderReconciliationRunPage:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProviderReconciliationRun"
        pagination:
          $ref: "#/components/schemas/Pagination"
...